small refactor

+++ src/common/utility.cpp
@@ -32,15 +32,15 @@ std::string load_file_into_string(const std::filesystem::path& filename)
 
 std::string_view trim_line(std::string_view sw_)
 {
-    const auto start = std::distance(sw_.begin(), std::find_if(sw_.begin(), sw_.end(), [](auto c) { return !std::isspace(c); }));
-    const auto end   = std::distance(sw_.rbegin(), std::find_if(sw_.rbegin(), sw_.rend(), [](auto c) { return !std::isspace(c); }));
+    const auto start = static_cast<std::size_t>(std::distance(sw_.begin(), std::find_if(sw_.begin(), sw_.end(), [](auto c) { return !std::isspace(c); })));
+    const auto end   = static_cast<std::size_t>(std::distance(sw_.rbegin(), std::find_if(sw_.rbegin(), sw_.rend(), [](auto c) { return !std::isspace(c); })));
     return sw_.substr(start, sw_.size() - (end + start));
 }
 
 std::string trim_line(std::string sw_)
 {
-    const auto start = std::distance(sw_.begin(), std::find_if(sw_.begin(), sw_.end(), [](auto c) { return !std::isspace(c); }));
-    const auto end   = std::distance(sw_.rbegin(), std::find_if(sw_.rbegin(), sw_.rend(), [](auto c) { return !std::isspace(c); }));
+    const auto start = static_cast<std::size_t>((std::distance(sw_.begin(), std::find_if(sw_.begin(), sw_.end(), [](auto c) { return !std::isspace(c); }))));
+    const auto end   = static_cast<std::size_t>(std::distance(sw_.rbegin(), std::find_if(sw_.rbegin(), sw_.rend(), [](auto c) { return !std::isspace(c); })));
     return sw_.substr(start, sw_.size() - (end + start));
 }
